+++ lib/access/resources/characteristic_values/cable_value.ex
@@ -24,20 +24,10 @@ defmodule DiffoExample.Access.CableValue do
 
     field :pairs, :integer, description: "the number of pairs in the cable"
 
-    field :length, :struct,
-      constraints: [instance_of: IntegerUnit],
-      description: "the length of the cable"
+    field :length, DiffoExample.Access.IntegerUnit, description: "the length of the cable"
 
-    field :loss, :struct,
-      constraints: [instance_of: FloatUnit],
-      description: "the loss of the cable at 300kHz"
+    field :loss, DiffoExample.Access.FloatUnit, description: "the loss of the cable at 300kHz"
 
     field :technology, :atom, description: "the cable technology"
-  end
-
-  defimpl String.Chars do
-    def to_string(struct) do
-      inspect(struct)
-    end
   end
 end